+++ client-react/src/pages/app/app-settings/FunctionRuntimeSettings/RuntimeVersionNew.tsx
@@ -1,0 +1,11 @@
+// TODO (krmitta):  Rename the file after this version is tested
+import React from 'react';
+import { AppSettingsFormProps } from '../AppSettings.types';
+
+const RuntimeVersion: React.FC<AppSettingsFormProps> = props => {
+  const { values } = props;
+
+  return <></>;
+};
+
+export default RuntimeVersion;
