+++ client-react/src/pages/app/app-settings/GeneralSettings/stacks/function-app/FunctionAppStackSettings.tsx
@@ -10,11 +10,7 @@ import { FunctionsRuntimeVersionHelper } from '../../../../../../utils/Functions
 import { SiteStateContext } from '../../../../../../SiteState';
 import { Field } from 'formik';
 import { isContainerApp } from '../../../../../../utils/arm-utils';
-import {
-  filterDeprecatedFunctionAppStack,
-  getStackVersionConfigPropertyName,
-  getStackVersionDropdownOptions,
-} from './FunctionAppStackSettings.data';
+import { filterDeprecatedFunctionAppStack, getStackVersionDropdownOptions } from './FunctionAppStackSettings.data';
 import { AppStackOs } from '../../../../../../models/stacks/app-stacks';
 import { settingsWrapper } from '../../../AppSettingsForm';
 import { Links } from '../../../../../../utils/FwLinks';
@@ -22,6 +18,7 @@ import TextField from '../../../../../../components/form-controls/TextField';
 import { IDropdownOption } from '@fluentui/react';
 import { AppSettingsFormValues, FormAppSetting } from '../../../AppSettings.types';
 import Dropdown from '../../../../../../components/form-controls/DropDown';
+import { getFunctionAppStackVersion, getStackVersionConfigPropertyName, isWindowsNodeApp } from '../../../../../../utils/stacks-utils';
 
 const FunctionAppStackSettings: React.FC<StackProps> = props => {
   const { t } = useTranslation();
@@ -45,21 +42,8 @@ const FunctionAppStackSettings: React.FC<StackProps> = props => {
     return initialValues.currentlySelectedStack;
   };
 
-  const isWindowsNodeApp = () => !isLinux() && runtimeStack && runtimeStack.toLowerCase() === WorkerRuntimeLanguages.nodejs;
-
   const getStackVersion = (values: AppSettingsFormValues, stack?: string) => {
-    if (isWindowsNodeApp()) {
-      const index = findFormAppSettingIndex([...values.appSettings], CommonConstants.AppSettingNames.websiteNodeDefaultVersion);
-      if (index !== -1) {
-        return values.appSettings[index].value;
-      } else {
-        return undefined;
-      }
-    } else {
-      const stackVersionProperty = getConfigProperty(stack);
-      const stackVersion = values.config && values.config && values.config.properties[stackVersionProperty];
-      return !!stackVersion ? stackVersion : undefined;
-    }
+    return getFunctionAppStackVersion(values, isLinux(), stack);
   };
 
   const getConfigProperty = (runtimeStack?: string) => {
@@ -107,7 +91,7 @@ const FunctionAppStackSettings: React.FC<StackProps> = props => {
   };
 
   const isVersionDirty = () => {
-    if (isWindowsNodeApp()) {
+    if (isWindowsNodeApp(isLinux(), runtimeStack)) {
       if (!!initialStackVersion && !!values.appSettings) {
         const index = findFormAppSettingIndex([...values.appSettings], CommonConstants.AppSettingNames.websiteNodeDefaultVersion);
         if (index !== -1) {
@@ -142,7 +126,7 @@ const FunctionAppStackSettings: React.FC<StackProps> = props => {
 
   const onMajorVersionChange = (_, option: IDropdownOption) => {
     setSelectedStackVersion(option.key as string);
-    if (isWindowsNodeApp()) {
+    if (isWindowsNodeApp(isLinux(), runtimeStack)) {
       const versionData = option.data;
       if (
         !!versionData &&
